refactor: Move `StdMutex` to its own header
This makes `threadsafety.h` agnostic to the actual implementation.
